FlyToTheSky/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.2.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
ca-certificates=2023.7.22=h56e8100_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.6.7=py38hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
filelock=3.12.3=pypi_0
fonttools=4.42.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib-resources=6.0.1=pypi_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.25.1=pyh6817e22_0
ipython=8.12.0=pyh08f2357_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
jupyter_client=8.3.1=pyhd8ed1ab_0
jupyter_core=5.3.1=py38haa244fe_0
kiwisolver=1.4.5=pypi_0
libffi=3.4.4=hd77b12b_0
libsodium=1.0.18=h8d14728_1
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
opencv-python=4.8.0.76=pypi_0
opencv-python-headless=4.8.0.76=pypi_0
openssl=3.0.10=h2bbff1b_2
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.3=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.5.0=pypi_0
pip=23.2.1=py38haa95532_0
platformdirs=3.10.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.5=pypi_0
pure_eval=0.2.2=pyhd8ed1ab_0
py-cpuinfo=9.0.0=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
python=3.8.17=h1aa4202_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2023.3=pypi_0
pywin32=305=py38h2bbff1b_0
pyyaml=6.0.1=pypi_0
pyzmq=23.2.1=py38h09162b1_0
requests=2.31.0=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
segment-anything=1.0=pypi_0
setuptools=68.0.0=py38haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
supervision=0.13.0=pypi_0
sympy=1.12=pypi_0
torch=2.0.1+cu117=pypi_0
torchaudio=2.0.2+cu117=pypi_0
torchvision=0.15.2+cu117=pypi_0
tornado=6.2=py38h294d835_0
tqdm=4.66.1=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023.3=pypi_0
ultralytics=8.0.166=pypi_0
urllib3=2.0.4=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.38.4=py38haa95532_0
zeromq=4.3.4=h0e60522_1
zipp=3.16.2=pyhd8ed1ab_0
